switched to using big decimal for better rounding and avoiding precision errors with binary representation of some double values

+++ src/main/java/com/example/calculator/transfer/incoming/BinaryOperationDTO.java
@@ -1,3 +1,5 @@
 package com.example.calculator.transfer.incoming;
 
-public record BinaryOperationDTO(Double operandOne, Double operandTwo, String operator) {}
+import java.math.BigDecimal;
+
+public record BinaryOperationDTO(BigDecimal operandOne, BigDecimal operandTwo, String operator) {}
